Make installAWSClient.sh better

+++ installAWSClient.sh
@@ -2,14 +2,26 @@
 # Copyright (c) 2011-2015 by Vertica, an HP Company.  All rights reserved.
 # Automates installation of AWS CLI
 
-sudo -E sh -c '(
+set -o errtrace
+set -o pipefail
+set -o errexit
+
 . ./autoscaling_vars.sh
+
 echo Install and configure AWS CLI
-cd /tmp
-yum install -y unzip
-curl https://s3.amazonaws.com/aws-cli/awscli-bundle.zip -o awscli-bundle.zip
-unzip -o awscli-bundle.zip
-./awscli-bundle/install -i /usr/local/aws -b /usr/local/bin/aws
+
+if [[ -f ~/.aws/config || -f ~/.aws/credentials ]]; then echo "not overwriting ~/.aws/config" ; exit 255; fi
+
+TEMP_AWSCLI=./tmp-awscli
+
+mkdir -p ${TEMP_AWSCLI}
+
+sh -c "(
+curl https://s3.amazonaws.com/aws-cli/awscli-bundle.zip -o ${TEMP_AWSCLI}/awscli-bundle.zip
+cd ${TEMP_AWSCLI} && unzip -o ./awscli-bundle.zip && cd -
+sudo -E ${TEMP_AWSCLI}/awscli-bundle/install -i /usr/local/aws -b /usr/local/bin/aws
+)" || rm -rf ${TEMP_AWSCLI}
+
 mkdir -p ~/.aws
 cat > ~/.aws/credentials <<EOF
 [default]
@@ -21,9 +33,7 @@ cat > ~/.aws/config <<EOF
 output = table
 region = $region
 EOF
-chmod 600 ~/.aws/*
-cp -R ~/.aws ~dbadmin
-chown -R dbadmin.verticadba ~dbadmin/.aws
-chmod 600 ~dbadmin/.aws/*
-)'
+
 echo AWS CLI installed
+
+rm -rf ${TEMP_AWSCLI}
